rfctr(html): drop parser parameter
HTML is parsed with `lxml.etree.HTMLParser`. After removing the
dependency on `XMLDocument` there is no need to specify the parser and
providing one other than `lxml.etree.HTMLParser` is likely to produce
unexpected results.

Note that it's not inconceivable we might want to add XHTML support and
for that we would likely use `lxml.etree.XMLParser`, but that is not
currently supported; there are no tests on XHTML documents and the
current behaviors are unlikely to be compatible without adjustments and
tests to defend those. Also, we would want to auto-detect XHTML rather
than require the user to specify a different parser, especially as a
Python class object.