rfctr(html): refine HTML parser
The initial version of the parser (purposely) raised on a block element
nested inside a phrasing element. While this is not valid according to
the HTML Standard, it is accepted by the browser and does happen in the
wild.

The refinements here handle this situation similarly to how the browser
does, breaking phrasing at the block element boundaries and starting it
up again after the block element.

Unfortunately this adds complexity to the parser, but it makes the
parser robust against pretty much any HTML we're likely to encounter and
partitions it consistent with how it would be rendered in the browser.