fix(file): temporary fix for FILETYPE_TO_MIMETYPE
`FILETYPE_TO_MIMETYPE` got wrong values for file-types accepting
multiple MIME-types because `MIMETYPES_TO_EXCLUDE` did not contain all
the values it needed to to produce a single canonical MIME-type for all
`FileType` members.

This is an expedient fix, but really we need to distinguish the single
canonical MIME-type for each `FileType` member, which gets written to
`.metadata.filetype` for that file-type, and the collection of
MIME-types and "aliases" which are recognized as that file-type.