rfctr(file): add new FileType enum
Elaborate the `FileType` enum to be a complete descriptor of file-types.
Add methods to allow `STR_TO_FILETYPE`, `EXT_TO_FILETYPE` and
`FILETYPE_TO_MIMETYPE` mappings to be replaced, removing those redundant
and noisy declarations.

This commit adds the new enum and its tests, but does not install it
yet. That will happen in a following commit.