rfctr(file): install _FileTypeDetectionContext
It's not much to look at yet, but this first step will enable the
remaining refactorings. So far it only extracts/encapsulates MIME-type
guessing and file-extension computation, but stay tuned for much more to
come.

Note this change pointed up a flaw in two main tests which did not
exercise the main detection algorithm but silently relied on filename
extension when the file did not exist at `file_path`.