rfctr(file): remove tests for obsolete behaviors
`detect_filetype()` now raises when called with a path that does not
correspond to a file on the filesystem. Confirmed with Matt this
detecting the filetype of a non-existent file (by using its extension
only) is not a desired or appropriate behavior for `detect_filetype()`.

Note the `FileType` corresponding to a particular filename extension can
easily be determined using `FileType.from_extension(ext)` without
engaging libmagic or any other file-type auto-detection.