rfctr(auto): distinct error for cannot partition
When `partition_email()` or `partition_msg()` partitions an attachment,
it needs to skip any attachment that cannot be partitioned, like maybe
an MP3 file or an SVG vector image.

We want `auto.partition()` to tell us when it doesn't know how to
partition a file rather than try to predict that using the file-type or
similar mechanism because any prediction would couple the EML and MSG
partitioners to auto-partitioning. This is undesirable because as we add
pluggable partitioners these modules have no good way to know which file
formats auto-partition can support.

Add and use `UnsupportedFileFormatError` in auto-partition when a
file-type cannot be partitioned because there is no partitioner mapped
to its file-type.

This allows clients like `partition_email()` to clearly distinguish
cases where an attachment should be skipped from partitioning errors
that should fail.