feat: chunk length ignores text_as_html
Prior to this change, the `.metadata.text_as_html` for a table was
considered when computing chunk length and deciding where to split a
table element. In practice, `.text_as_html` is always longer than text
so it was the primary criteria for locating splits.

Remove that behavior and compute chunk length solely on `element.text`.